feat: Implementing the Todo Get and Post Views
Adding the base logic for the Todo feature. For now three views
are implemented:

- GET /api/v1/todos: This view will fetch for all the created views
in the database and return them to the user

- GET /api/v1/todos/imdbId: This view will fetch for the Todo
coresponding to the imdbId value and return it to the user if found.
This view doesn't handle the case where the Todo isn't found.

- POST /api/v1/todos/: This view will take a JSON and use the data
to create a new Todo item.