feat: Adding the Todo update and delete views
The Todo update and delete views where successfully implemented
and tested, the two views are accessible by the following endpoints:

- POST /api/v1/todos/imdbId: Will fetch for the Todo with the corresponding
imdbId values then updates it with the new title and body values. Note
that this view will always update both title and body, even if one or
both values aren't given, which will eventually lead to a null value
being assigned to the missing values. Will return the message "This Todo
doesn't exist" if there is no Todo with that imdbId.

- DELETE /api/v1/todos/imdbId: will fetch for the Todo with the
corresponding imdbId and deletes it from the database. Will return
the message "This Todo doesn't exist" if there is no Todo with that
imdbId.